Add tests, commandline and docs for include_files and exclude_files

+++ tests/test_cmdline.py
@@ -17,15 +17,21 @@ class TestCommandLineBasic(unittest.TestCase):
 
     def cleanup(self, output_file="result.html"):
         os.remove("cmdline_test.py")
-        os.remove(output_file)
+        if output_file:
+            os.remove(output_file)
 
     def template(self, cmd_list, expected_output_file="result.html", success=True):
         self.build_script()
         result = subprocess.run(cmd_list, stdout=subprocess.PIPE)
         self.assertTrue(success ^ (result.returncode != 0))
-        self.assertTrue(os.path.exists(expected_output_file))
+        if expected_output_file:
+            self.assertTrue(os.path.exists(expected_output_file))
         self.cleanup(output_file=expected_output_file)
         return result
+
+    def test_no_file(self):
+        result = self.template(["python", "-m", "viztracer"], expected_output_file=None)
+        self.assertIn("help", result.stdout.decode("utf8"))
 
     def test_run(self):
         self.template(["python", "-m", "viztracer", "cmdline_test.py"])
@@ -48,4 +54,16 @@ class TestCommandLineBasic(unittest.TestCase):
 
     def test_max_stack_depth(self):
         self.template(["python", "-m", "viztracer", "--max_stack_depth", "5", "cmdline_test.py"])
+    
+    def test_include_files(self):
+        result = self.template(["python", "-m", "viztracer", "--include_files", "./abcd", "cmdline_test.py"], expected_output_file=None)
+        self.assertIn("help", result.stdout.decode("utf8"))
+        self.template(["python", "-m", "viztracer", "--include_files", "./", "--run", "cmdline_test.py"])
+        self.template(["python", "-m", "viztracer", "--include_files", "./", "--max_stack_depth", "5", "cmdline_test.py"])
+        self.template(["python", "-m", "viztracer", "--include_files", "./abcd", "--run", "cmdline_test.py"])
 
+    def test_exclude_files(self):
+        result = self.template(["python", "-m", "viztracer", "--exclude_files", "./abcd", "cmdline_test.py"], expected_output_file=None)
+        self.assertIn("help", result.stdout.decode("utf8"))
+        self.template(["python", "-m", "viztracer", "--exclude_files", "./", "--run", "cmdline_test.py"])
+        self.template(["python", "-m", "viztracer", "--exclude_files", "./abcd", "--run", "cmdline_test.py"])
